Add mixed precision and 16-bit support to `TorchForecastingModel`
Previously, the `"precision"` option set in `pl_trainer_kwargs` is
ignored in favour of precision of the time series data type. That led to
the lack of support for mixed precision training and 16-bit true
precison despite `pytorch-lightning.Trainer` native support.

I added support for mixed precision and any other preicison options
supported by Lightning in `TorchForecastingModel`.

Here, by default when the `"precision"` option is not set in
`pl_trainer_kwargs` or a custom trainer, the model precision is
determined by the data type. When it is set, we can assume that the user
understand the ramifications of that option and the option is always
passed along to the trainer, even when there is a mismatch between the
model precision and the data type.

Sometimes the mismatch is necessary when the data type is 32-bit but the
model can be trained in mixed precision. The other times the mismatch
would cause an error, say, training 32-bit model on 64-bit data.

**Precision is an advanced option for `TorchForecastingModel` training and
users should only use it at their own benefits and risks!**