Bugfix: Recommend `pytorch-lightning>=2.5.3` for Google Colab notebook
Fixes #2670.

Some versions of `pytorch-lightning` may not be comptaible with
`torch_vla>=2.7.0` anymore due to deprecation of XLA APIs since 2.7.
This was fixed in [Lightning
2.5.3](https://github.com/Lightning-AI/pytorch-lightning/releases/tag/2.5.3).
See Lightning's
[#20852](https://github.com/Lightning-AI/pytorch-lightning/issues/20852)
for the upstream bug fix.

However, running the "21-TSMixer-examples", the notebook never got past
"Model Training". Using the latest nightly build XLA 2.9.0 fixes the
issue, as instructed on [pytorch/xla](https://github.com/pytorch/xla).
The exact reasons why that is are unknown.