Add mixed precision and 16-bit support to `TorchForecastingModel` (#2883)
* Add mixed precision and 16-bit support to `TorchForecastingModel`

Previously, the `"precision"` option set in `pl_trainer_kwargs` is
ignored in favour of precision of the time series data type. That led to
the lack of support for mixed precision training and 16-bit true
precison despite `pytorch-lightning.Trainer` native support.

I added support for mixed precision and any other preicison options
supported by Lightning in `TorchForecastingModel`.

Here, by default when the `"precision"` option is not set in
`pl_trainer_kwargs` or a custom trainer, the model precision is
determined by the data type. When it is set, we can assume that the user
understand the ramifications of that option and the option is always
passed along to the trainer, even when there is a mismatch between the
model precision and the data type.

Sometimes the mismatch is necessary when the data type is 32-bit but the
model can be trained in mixed precision. The other times the mismatch
would cause an error, say, training 32-bit model on 64-bit data.

**Precision is an advanced option for `TorchForecastingModel` training and
users should only use it at their own benefits and risks!**

* Update CHANGELOG with mixed & 16-true precision support

* Fix test fails due to Python 3.13 error hint

Two tests failed on Python 3.13 due to NEW keyword argument suggestion
in Python 3.13. See [What’s New In Python
3.13](https://docs.python.org/3/whatsnew/3.13.html) and
[gh-107944](https://github.com/python/cpython/issues/107944) for
details.

Those fails are unrelated to the new precision option support.

* Fix a bug when `"bf16-mixed"` precision is used

While there are many mixed precision options on GPU, the only mixed
precision on CPU currently is `"bf16-mixed"`. The model output would be
"BFloat16" and numpy does not support conversion from it. We must
convert the output to float32 first.

* Add two tests of preicison option for better code coverage

1. Add `test_auto_precision_casting` to test auto model preicision
   casting based on time series data type, when precision is not set.
2. Add `test_mixed_precision_training` to test mixed precision training
   when precision option is `"16"` (`"16-mixed"` on Lightning>=2.0.0).

* update changelog

* remove redundant diffs

* add back diffs for testing with python 3.13

* Produce warnings for 16-bit time series and user-defined precision

* Add mixed-precision tests & adjust 16-bit tests

1. Add tests for both mixed preicison options, i.e., "16-mixed" and
   "bf16-mixed:.
2. Adjust tests when inputs are 16-bit to allow for 32-bit predictions.

* Apply suggestions from code review

Log to into for user-defined precision and to warning when float16-like precision is being used.

Co-authored-by: Dennis Bader <dennis.bader@gmx.ch>

* Test valid precision options for custom & built-in trainers

Also ensure there are no NaN values in predictions.

---------

Co-authored-by: dennisbader <dennis.bader@gmx.ch>